Some cleanup. Switched LTTNG and PerfData to use new common EventProjection
