+++ LTTngDataExtensions/LTTngDataExtensions.csproj
@@ -55,6 +55,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\LTTngCds\LTTngCds.csproj" />
+    <ProjectReference Include="..\Utilities\Utilities.csproj" />
   </ItemGroup>
 
   <ItemGroup>
