+++ PerfDataExtensions/Tables/GenericEventTable.cs
@@ -2,13 +2,12 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Diagnostics.CodeAnalysis;
-using System.Security.Cryptography;
 using CtfPlayback.FieldValues;
 using PerfDataExtensions.DataOutputTypes;
 using Microsoft.Performance.SDK;
 using Microsoft.Performance.SDK.Extensibility;
 using Microsoft.Performance.SDK.Processing;
+using Utilities;
 
 namespace PerfDataExtensions.Tables
 {
@@ -74,7 +73,7 @@ namespace PerfDataExtensions.Tables
 
             var tableGenerator = tableBuilder.SetRowCount((int)events.Count);
 
-            var genericEventProjection = new GenericEventProjection(events);
+            var genericEventProjection = new EventProjection<PerfGenericEvent>(events);
 
             var eventNameColumn = new BaseDataColumn<string>(
                 eventNameColumnConfig, 
@@ -110,7 +109,7 @@ namespace PerfDataExtensions.Tables
 
                 // generate a column configuration
                 var fieldColumnConfiguration = new ColumnConfiguration(
-                    new ColumnMetadata(GenerateGuidFromName(fieldName), fieldName, genericEventFieldNameProjection, fieldName),
+                    new ColumnMetadata(Common.GenerateGuidFromName(fieldName), fieldName, genericEventFieldNameProjection, fieldName),
                     new UIHints
                     {
                         IsVisible = true,
@@ -124,67 +123,6 @@ namespace PerfDataExtensions.Tables
                 tableGenerator.AddColumn(fieldColumnConfiguration, genericEventFieldAsStringProjection);
             }
         }
-
-        [SuppressMessage("Microsoft.Security.Cryptography", "CA5354:SHA1CannotBeUsed", Justification = "Not a security related usage - just generating probabilistically unique id to identify a column from its name.")]
-        private static Guid GenerateGuidFromName(string name)
-        {
-            // The algorithm below is following the guidance of http://www.ietf.org/rfc/rfc4122.txt
-            // Create a blob containing a 16 byte number representing the namespace
-            // followed by the unicode bytes in the name.  
-            var bytes = new byte[name.Length * 2 + 16];
-            uint namespace1 = 0x482C2DB2;
-            uint namespace2 = 0xC39047c8;
-            uint namespace3 = 0x87F81A15;
-            uint namespace4 = 0xBFC130FB;
-            // Write the bytes most-significant byte first.  
-            for (int i = 3; 0 <= i; --i)
-            {
-                bytes[i] = (byte)namespace1;
-                namespace1 >>= 8;
-                bytes[i + 4] = (byte)namespace2;
-                namespace2 >>= 8;
-                bytes[i + 8] = (byte)namespace3;
-                namespace3 >>= 8;
-                bytes[i + 12] = (byte)namespace4;
-                namespace4 >>= 8;
-            }
-            // Write out  the name, most significant byte first
-            for (int i = 0; i < name.Length; i++)
-            {
-                bytes[2 * i + 16 + 1] = (byte)name[i];
-                bytes[2 * i + 16] = (byte)(name[i] >> 8);
-            }
-
-            // Compute the Sha1 hash 
-            var sha1 = SHA1.Create();
-            byte[] hash = sha1.ComputeHash(bytes);
-
-            // Create a GUID out of the first 16 bytes of the hash (SHA-1 create a 20 byte hash)
-            int a = (((((hash[3] << 8) + hash[2]) << 8) + hash[1]) << 8) + hash[0];
-            short b = (short)((hash[5] << 8) + hash[4]);
-            short c = (short)((hash[7] << 8) + hash[6]);
-
-            c = (short)((c & 0x0FFF) | 0x5000);   // Set high 4 bits of octet 7 to 5, as per RFC 4122
-            Guid guid = new Guid(a, b, c, hash[8], hash[9], hash[10], hash[11], hash[12], hash[13], hash[14], hash[15]);
-            return guid;
-        }
-    }
-
-    public struct GenericEventProjection
-        : IProjection<int, PerfGenericEvent>
-    {
-        private readonly ProcessedEventData<PerfGenericEvent> genericEvents;
-
-        public GenericEventProjection(ProcessedEventData<PerfGenericEvent> genericEvents)
-        {
-            this.genericEvents = genericEvents;
-        }
-
-        public Type SourceType => typeof(int);
-
-        public Type ResultType => typeof(PerfGenericEvent);
-
-        public PerfGenericEvent this[int value] => this.genericEvents[(uint)value];
     }
 
     public struct GenericEventFieldProjection
